fix(dashboard): code review: sequential screenshot, annotate.spec.ts
